+++ src/SDK/Client/Platform/methods/identities/register.ts
@@ -1,7 +1,8 @@
 import { Platform } from "../../Platform";
 import { wait } from "../../../../../utils/wait";
-import { createFakeInstantLock } from "../../../../../utils/createFakeIntantLock";
 import createAssetLockTransaction from "../../createAssetLockTransaction";
+import createIdentityTransition from "./internal/createIdentityTransition";
+import broadcastIdentityCreateStateTransition from "./internal/broadcastIdentityCreateTransition";
 
 /**
  * Register identities to the platform
@@ -13,7 +14,7 @@ export default async function register(
   this: Platform,
   fundingAmount : number = 10000
 ): Promise<any> {
-    const { client, dpp, passFakeAssetLockProofForTests } = this;
+    const { client } = this;
 
     const account = await client.getWalletAccount();
 
@@ -29,50 +30,11 @@ export default async function register(
     // Wait some time for propagation
     await wait(1000);
 
-    const identityIndex = await account.getUnusedIdentityIndex();
-
-    // @ts-ignore
-    const { privateKey: identityPrivateKey } = account.getIdentityHDKeyByIndex(identityIndex, 0);
-    const identityPublicKey = identityPrivateKey.toPublicKey();
-
-    let instantLock;
-    // Create poof that the transaction won't be double spend
-    if (passFakeAssetLockProofForTests) {
-        instantLock = createFakeInstantLock(assetLockTransaction.hash);
-    } else {
-        instantLock = await account.waitForInstantLock(assetLockTransaction.hash);
-    }
-    // @ts-ignore
-    const assetLockProof = await dpp.identity.createInstantAssetLockProof(instantLock);
-
-    // Create Identity
-    // @ts-ignore
-    const identity = dpp.identity.create(
-        assetLockTransaction, assetLockOutputIndex, assetLockProof, [identityPublicKey]
+    const { identity, identityCreateTransition, identityIndex } = await createIdentityTransition(
+        this, assetLockTransaction, assetLockOutputIndex, assetLockPrivateKey
     );
 
-    // Create ST
-    const identityCreateTransition = dpp.identity.createIdentityCreateTransition(identity);
-
-    identityCreateTransition.signByPrivateKey(assetLockPrivateKey);
-
-    const result = await dpp.stateTransition.validateStructure(identityCreateTransition);
-
-    if (!result.isValid()) {
-        throw new Error(`StateTransition is invalid - ${JSON.stringify(result.getErrors())}`);
-    }
-
-    // Broadcast ST
-    await client.getDAPIClient().platform.broadcastStateTransition(identityCreateTransition.toBuffer());
-
-    account.storage.insertIdentityIdAtIndex(
-        account.walletId,
-        identity.getId().toString(),
-        identityIndex,
-    );
-
-    // Wait some time for propagation
-    await wait(6000);
+    await broadcastIdentityCreateStateTransition(this, identityCreateTransition, identity, identityIndex);
 
     let fetchedIdentity;
     do {
